requirements.txt
# Core AI Dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0

# Optional GPU Optimizations
# triton>=2.1.0  # Not available for Python 3.13 on Windows yet

# Game Engine
pygame>=2.5.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Optional Upscaling
# realesrgan

# Utilities
requests>=2.31.0
tqdm>=4.66.0
safetensors>=0.4.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
